fix: close milestone 40.4 review pass 3
