+++ plans/reviews/archive/phase-40-milestone-40-4-claude-opus-review-pass-3.md
@@ -1,0 +1,38 @@
+I re-ran the milestone's gates and audited the full diff read-only. Working tree and both submodules are clean; pointer is `a980835e6→d7577d492` with `vscode = 273fd5d3e`, matching `vscode_extension_rules.json:22-23`.
+
+## Checks re-run
+
+| Check | Result |
+|---|---|
+| `documentation --suite structure --suite ga-release` | PASS, variants=2, failures=0 |
+| `developer_tooling --suite editor-release` | PASS, variants=6, failures=0 |
+| `distribution_release --suite qualification --suite full` | PASS, variants=55, failures=0 |
+| `check_ga_release_docs.py --self-test` | PASS (13 mutation cases incl. `global-preview-claim`) |
+| `check_file_size_guardrails.py` | PASS (2879 files, limit 900) |
+| Scope | 0 files under `crates/**`; no Marketplace publication surface in `.github/workflows/` |
+
+## Pass-2 observations — all closed
+
+Fixture headroom (`qualification_fixture.py` 880→828, `qualification_selftest.py` 819, new `qualification_rust_fixture.py`); semantic `FORBIDDEN_CLAIM_PATTERNS` at `check_ga_release_docs.py:140-146` with the self-test mutation reworded to *"Sifr remains in public preview"* (`:413`); `planner.py:12→25` import order corrected; the upstream package/pointer ledger and both main-review entries are recorded at `plans/issues/active/phase-40-stable-channel-ga-execution.md:120-154`.
+
+Pass-1 closures also verified: all-docs sweep (`load_public_documents`, `:178-192`), truthful `status: "planned"` Marketplace plan (`qualify_stable_editor.py:335-352`, asserted at `editor_qualification.py:95-138`), `rollback_version` dispatch input, public `sifr emit` limitation on all four advertising pages, `shlex.join`/`shlex.split`, forbidden Windows triples, appending `unsupported-target-claim` mutation.
+
+## Findings
+
+**1. HIGH — the new `rollback_version` plumbing lands in the wrong job, so the `editor` qualification job fails on every dispatch.** `.github/workflows/release-qualification.yml:200` passes `--rollback-version "${ROLLBACK_VERSION}"`, but that step's `env:` block (`:176-178`) declares only `SOURCE_COMMIT` and `VERSION`. The variable is not defined at workflow or job level (only occurrences are `:40`, `:119`, `:200`), and the step runs `set -euo pipefail` (`:180`). Under `-u` the expansion aborts the step with `ROLLBACK_VERSION: unbound variable` — after `npm ci`, the five npm scripts, and the artifact download have already run. Conversely `:119` adds `ROLLBACK_VERSION` to the `build` job step, which never references it: the env line was attached to the wrong step. Net effect: `qualification-editor.json` is never produced, so `collect`/`assemble` cannot bind editor evidence and no stable candidate can be qualified — first-GA or otherwise. The pass-1 finding-3 remediation is therefore unrealized in production; it is proven only by fixture and by the demo, which passes `--rollback-version none` literally (`demos/editor_candidate_qualification_demo.sh:78`). The immutable-contract case does not catch it: `release_qualification_workflow_contract.sh:80` greps for the literal argument text and `:69` for the `validate` job's `GITHUB_OUTPUT` line, but nothing asserts the `editor` step's env binding. *Remediation:* move `ROLLBACK_VERSION: ${{ needs.validate.outputs.rollback_version }}` from `:119` into the editor step's `env:` at `:176-178`, and extend the contract case to assert that every `${VAR}` referenced in a step's `run:` is declared in that step's `env:` (or at minimum add `ROLLBACK_VERSION: ${{ needs.validate.outputs.rollback_version }}` to the required-text tuple scoped to the editor job).
+
+**2. MEDIUM — GA docs instruct users to install the extension from a Marketplace listing this milestone deliberately does not publish.** `docs/cli/lsp.mdx:68`: *"Install the Sifr extension from the VS Code Marketplace."* The whole point of 40_4 is *"Qualify the exact VSIX and Marketplace identity **without publication**"*; `qualify_stable_editor.py:351` records `status: "planned"` and the phase DoD now defers the credentialed publish to `milestone_40_5` (`plans/phases/40_stable_channel_ga_promotion_and_release_governance.md:904-907`). At `0.1.0` GA this instruction sends users to a listing that does not exist, and it contradicts `docs/releases/compatibility.mdx:20-25`, which describes `sifr.sifr-vscode` `0.2.0` only as *qualified*. The new all-docs sweep structurally cannot catch it — `FORBIDDEN_CLAIMS`/`FORBIDDEN_CLAIM_PATTERNS` (`check_ga_release_docs.py:131-146`) have no entry for Marketplace availability, and `lsp.mdx` is outside `CANONICAL_DOCUMENTS` so no required-fact assertion covers it either. *Remediation:* reword `lsp.mdx:68` to reflect the actual `0.1.0` install path (packaged VSIX / `sifr.serverPath`), and add a forbidden pattern for Marketplace-availability phrasing so 40_5 must flip it deliberately when publication lands.
+
+**3. LOW/MEDIUM — the milestone demo does not exercise the `tests` step its own DoD claims.** `plans/phases/40_stable_channel_ga_promotion_and_release_governance.md:932-935` states the demo will *"exercise linting, check, and **tests** through the exact candidate compiler."* `demos/editor_candidate_qualification_demo.sh:92-98` runs only `check`, `fmt`, `fmt --check`, and `lint` against the candidate; there is no `"${candidate_binary}" test` invocation anywhere in the file (the `test` tokens at `:65`, `:73` are shell builtins). Since `docs/releases/0.1.0.mdx:20` advertises *"compiler-backed check, lint, and test actions"* as a GA highlight, the test path is advertised but unproven by the packaged candidate. *Remediation:* add a `sifr test` invocation over a fixture directory containing a test function inside the `:92-98` block, or narrow the DoD demo text to what is actually exercised.
+
+## Non-actionable observations
+
+- `verification/areas/distribution_release/governance/selftest.py` is now 884/900 (grew 11 this round) — the closest file to the cap after the fixture split; next refactor candidate.
+- `FORBIDDEN_CLAIM_PATTERNS` covers only `*-pc-windows-msvc`; an unsupported triple such as `aarch64-unknown-linux-musl` would still pass. A positive allowlist of the four `TARGETS` would close the class.
+- `qualify_stable_documentation.py` still has no production caller — only `--self-test` from `stable_editor_qualification_contract.sh:8`; the report it emits is bound by the planner but nothing tells an operator to run it.
+- `stable_site_release_facts.json` remains a synthetic fixture (`aaaa…`/`ffff…` digests); the real plan/index binding arrives with the 40_5 site adapter.
+- `vsix_install_smoke` is archive extraction plus a `config.js` `serverCommand()` assertion (`qualify_stable_editor.py:275-288`), not a VS Code install; the field name still overstates it.
+- The demo requires a clean checkout (`:13-16`) then dirties the submodule via `npm ci`/`npm run package`; benign because `dist/` and `node_modules/` are ignored.
+- All `milestone_40_4` checkboxes correctly remain unchecked, including candidate-evidence materialization — the milestone cannot close on this diff alone.
+
+Findings 1–3 are actionable, so this round is not approved.
